# Conflicts:
#
BabySleep.xcodeproj/project.xcworkspace/xcuserdata/medica.xcuserdatad/Us
erInterfaceState.xcuserstate

+++ BabySleep/MMPDeepSleepPreventer/MMPDeepSleepPreventer.m
@@ -208,7 +208,7 @@
         
         if (propertySetError)
         {
-            MMPALog(@"Error setting kAudioSessionProperty_OverrideCategoryMixWithOthers: %d", propertySetError);
+            MMPALog(@"Error setting kAudioSessionProperty_OverrideCategoryMixWithOthers: %d", (int)propertySetError);
         }
         
 #pragma clang diagnostic pop
